Quell undesired operation of more hook in post->content_out.
The post "desolation" in local test env would always display the content_out --More-- tag on the index view. We want the Content_out to only do anything when a user has typed a <--more--> tag into the editor for the post.

Apply_with_hook_params() takes 5 arguments, the last two of which are an integer number of characters and an integer number of paragraphs. Specifying nothing for those final two arguments instead of placeholding with zeroes stops the more link being output for the test post on the index view.